+++ contracts/registry/src/error.rs
@@ -1,11 +1,10 @@
-use loam_sdk::soroban_sdk::{self, contracterror};
+use soroban_sdk::{self, contracterror};
 
 #[contracterror]
 #[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
 #[repr(u32)]
 pub enum Error {
-    /// No such Contract has been published
-    NoSuchContractPublished = 1,
+    NoSuchWasmPublished = 1,
     /// No such version of the contact has been published
     NoSuchVersion = 2,
     /// Wasm name already claimed
@@ -21,7 +20,7 @@ pub enum Error {
     /// New version must be greater than the most recent version
     VersionMustBeGreaterThanCurrent = 8,
     /// Invalid name.
-    /// Must be 64 characters or less;
+    /// Must be at most 64 characters and non-empty;
     /// ascii alphanumeric, '-', or '_';
     /// start with a ascii alphabetic character;
     /// and not be a Rust keyword