+++ contracts/registry/src/util.rs
@@ -1,7 +1,6 @@
-use loam_sdk::soroban_sdk::{crypto::Hash, env, Bytes, String};
+use soroban_sdk::{crypto::Hash, Bytes, Env, String};
 
-pub fn hash_string(s: &String) -> Hash<32> {
-    let env = env();
+pub fn hash_string(env: &Env, s: &String) -> Hash<32> {
     let len = s.len() as usize;
     let mut bytes = [0u8; 100];
     let bytes = &mut bytes[0..len];
@@ -12,5 +11,3 @@ pub fn hash_string(s: &String) -> Hash<32> {
 }
 
 pub const MAX_BUMP: u32 = 535_679;
-
-pub(crate) const REGISTRY: &str = "registry";