+++ crates/stellar-build/CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5](https://github.com/theahaco/scaffold-stellar/compare/stellar-build-v0.0.4...stellar-build-v0.0.5) - 2025-11-10
+
+### Other
+
+- update Cargo.toml dependencies
+
+## [0.0.4](https://github.com/theahaco/scaffold-stellar/compare/stellar-build-v0.0.3...stellar-build-v0.0.4) - 2025-10-20
+
+### Added
+
+- add network in target wasm path ([#213](https://github.com/theahaco/scaffold-stellar/pull/213))
+
+### Other
+
+- Add macro for `stellar_asset` ([#214](https://github.com/theahaco/scaffold-stellar/pull/214))
+
 ## [0.0.3](https://github.com/theahaco/scaffold-stellar/compare/stellar-build-v0.0.2...stellar-build-v0.0.3) - 2025-09-26
 
 ### Fixed