+++ crates/stellar-scaffold-macro/CHANGELOG.md
@@ -7,6 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.10](https://github.com/theahaco/scaffold-stellar/compare/stellar-scaffold-macro-v0.8.9...stellar-scaffold-macro-v0.8.10) - 2025-11-10
+
+### Other
+
+- *(deps)* bump the cargo-minor group across 1 directory with 17 updates ([#284](https://github.com/theahaco/scaffold-stellar/pull/284))
+
+## [0.8.9](https://github.com/theahaco/scaffold-stellar/compare/stellar-scaffold-macro-v0.8.8...stellar-scaffold-macro-v0.8.9) - 2025-10-20
+
+### Added
+
+- export both admit and normal interface ([#218](https://github.com/theahaco/scaffold-stellar/pull/218))
+
+### Fixed
+
+- *(stellar-registry)* stellar_asset -> import_asset ([#215](https://github.com/theahaco/scaffold-stellar/pull/215))
+
+### Other
+
+- Add macro for `stellar_asset` ([#214](https://github.com/theahaco/scaffold-stellar/pull/214))
+
 ## [0.8.8](https://github.com/theahaco/scaffold-stellar/compare/stellar-scaffold-macro-v0.8.7...stellar-scaffold-macro-v0.8.8) - 2025-09-26
 
 ### Fixed